# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MALI-METEO Bulletins API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SituationGenerale {
  ciel: string;
  vents: string;
  visibilite: string;
  /** @nullable */
  orages?: string | null;
  temperatures: string;
}

/**
 * @nullable
 */
export type VilleDataCondition = typeof VilleDataCondition[keyof typeof VilleDataCondition] | null;


export const VilleDataCondition = {
  ensoleille: 'ensoleille',
  nuageux: 'nuageux',
  partiellement_nuageux: 'partiellement_nuageux',
  pluvieux: 'pluvieux',
  orageux: 'orageux',
  couvert: 'couvert',
} as const;

export interface VilleData {
  nom: string;
  /** @nullable */
  tmax?: number | null;
  /** @nullable */
  tmin?: number | null;
  /** @nullable */
  directionVent?: string | null;
  /** @nullable */
  vitesseVent?: number | null;
  /** @nullable */
  condition?: VilleDataCondition;
}

export type VigilanceDataNiveau = typeof VigilanceDataNiveau[keyof typeof VigilanceDataNiveau];


export const VigilanceDataNiveau = {
  vigilance_absolue: 'vigilance_absolue',
  tres_vigilants: 'tres_vigilants',
  attentifs: 'attentifs',
  pas_vigilance: 'pas_vigilance',
} as const;

export interface VigilanceData {
  region: string;
  niveau: VigilanceDataNiveau;
}

export type BulletinType = typeof BulletinType[keyof typeof BulletinType];


export const BulletinType = {
  radio: 'radio',
  matinal: 'matinal',
  journaux: 'journaux',
  ortm: 'ortm',
  national: 'national',
} as const;

export interface Bulletin {
  id: number;
  type: BulletinType;
  bulletinDate: string;
  periodLabel: string;
  validiteLabel: string;
  /** @nullable */
  heureLabel?: string | null;
  situationGenerale: SituationGenerale;
  donneesVilles: VilleData[];
  vigilanceNiveaux: VigilanceData[];
  createdAt: string;
  updatedAt: string;
}

export type BulletinInputType = typeof BulletinInputType[keyof typeof BulletinInputType];


export const BulletinInputType = {
  radio: 'radio',
  matinal: 'matinal',
  journaux: 'journaux',
  ortm: 'ortm',
  national: 'national',
} as const;

export interface BulletinInput {
  type: BulletinInputType;
  bulletinDate: string;
  periodLabel: string;
  validiteLabel: string;
  /** @nullable */
  heureLabel?: string | null;
  situationGenerale: SituationGenerale;
  donneesVilles: VilleData[];
  vigilanceNiveaux?: VigilanceData[];
}

export type BulletinStatsParType = {[key: string]: number};

export interface BulletinStats {
  total: number;
  parType: BulletinStatsParType;
  recents: Bulletin[];
}

export type TemplateCategorie = typeof TemplateCategorie[keyof typeof TemplateCategorie];


export const TemplateCategorie = {
  ciel: 'ciel',
  vents: 'vents',
  visibilite: 'visibilite',
  orages: 'orages',
  temperatures: 'temperatures',
} as const;

export interface Template {
  id: number;
  nom: string;
  categorie: TemplateCategorie;
  texte: string;
  estDefaut: boolean;
  createdAt: string;
}

export type TemplateInputCategorie = typeof TemplateInputCategorie[keyof typeof TemplateInputCategorie];


export const TemplateInputCategorie = {
  ciel: 'ciel',
  vents: 'vents',
  visibilite: 'visibilite',
  orages: 'orages',
  temperatures: 'temperatures',
} as const;

export interface TemplateInput {
  nom: string;
  categorie: TemplateInputCategorie;
  texte: string;
}

export type ListBulletinsParams = {
type?: ListBulletinsType;
limit?: number;
offset?: number;
};

export type ListBulletinsType = typeof ListBulletinsType[keyof typeof ListBulletinsType];


export const ListBulletinsType = {
  radio: 'radio',
  matinal: 'matinal',
  journaux: 'journaux',
  ortm: 'ortm',
  national: 'national',
} as const;

export type ListTemplatesParams = {
categorie?: ListTemplatesCategorie;
};

export type ListTemplatesCategorie = typeof ListTemplatesCategorie[keyof typeof ListTemplatesCategorie];


export const ListTemplatesCategorie = {
  ciel: 'ciel',
  vents: 'vents',
  visibilite: 'visibilite',
  orages: 'orages',
  temperatures: 'temperatures',
} as const;