# lib/api-zod/src/generated/types/bulletin.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MALI-METEO Bulletins API
 * OpenAPI spec version: 0.1.0
 */
import type { BulletinType } from './bulletinType';
import type { SituationGenerale } from './situationGenerale';
import type { VigilanceData } from './vigilanceData';
import type { VilleData } from './villeData';

export interface Bulletin {
  id: number;
  type: BulletinType;
  bulletinDate: string;
  periodLabel: string;
  validiteLabel: string;
  /** @nullable */
  heureLabel?: string | null;
  situationGenerale: SituationGenerale;
  donneesVilles: VilleData[];
  vigilanceNiveaux: VigilanceData[];
  createdAt: string;
  updatedAt: string;
}