# lib/api-zod/src/generated/types/bulletinStats.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MALI-METEO Bulletins API
 * OpenAPI spec version: 0.1.0
 */
import type { Bulletin } from './bulletin';
import type { BulletinStatsParType } from './bulletinStatsParType';

export interface BulletinStats {
  total: number;
  parType: BulletinStatsParType;
  recents: Bulletin[];
}