# lib/api-zod/src/generated/types/situationGenerale.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MALI-METEO Bulletins API
 * OpenAPI spec version: 0.1.0
 */

export interface SituationGenerale {
  ciel: string;
  vents: string;
  visibilite: string;
  /** @nullable */
  orages?: string | null;
  temperatures: string;
}