# lib/api-zod/src/generated/types/template.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MALI-METEO Bulletins API
 * OpenAPI spec version: 0.1.0
 */
import type { TemplateCategorie } from './templateCategorie';

export interface Template {
  id: number;
  nom: string;
  categorie: TemplateCategorie;
  texte: string;
  estDefaut: boolean;
  createdAt: string;
}