# lib/api-zod/src/generated/types/templateInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MALI-METEO Bulletins API
 * OpenAPI spec version: 0.1.0
 */
import type { TemplateInputCategorie } from './templateInputCategorie';

export interface TemplateInput {
  nom: string;
  categorie: TemplateInputCategorie;
  texte: string;
}