# lib/api-zod/src/generated/types/vigilanceData.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MALI-METEO Bulletins API
 * OpenAPI spec version: 0.1.0
 */
import type { VigilanceDataNiveau } from './vigilanceDataNiveau';

export interface VigilanceData {
  region: string;
  niveau: VigilanceDataNiveau;
}