# lib/api-zod/src/generated/types/villeData.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MALI-METEO Bulletins API
 * OpenAPI spec version: 0.1.0
 */
import type { VilleDataCondition } from './villeDataCondition';

export interface VilleData {
  nom: string;
  /** @nullable */
  tmax?: number | null;
  /** @nullable */
  tmin?: number | null;
  /** @nullable */
  directionVent?: string | null;
  /** @nullable */
  vitesseVent?: number | null;
  /** @nullable */
  condition?: VilleDataCondition;
}